Adding pre-research material for EMIntegration and moved Database endpoints to subfolder
