+++ plugins/verticalapp-driver/src/Api/Database/post_meta.php
@@ -1,8 +1,8 @@
 <?php
 
-namespace VerticalAppDriver\Api;
+namespace VerticalAppDriver\Api\Database;
 
-require_once __DIR__ . '/../Auth/apikey_checking.php';
+require_once __DIR__ . '/../../Auth/apikey_checking.php';
 
 use WP_REST_Request;
