+++ plugins/verticalapp-driver/verticalapp-driver.php
@@ -14,12 +14,12 @@ if (!defined('ABSPATH')) exit;
 // Composer autoloader
 require_once __DIR__ . '/vendor/autoload.php';
 require_once __DIR__ . '/src/constants.php'; // adjust path if needed
-require_once __DIR__ . '/src/Api/routes.php';
+require_once __DIR__ . '/src/Api/Database/routes.php';
 require_once __DIR__ . '/src/Helpers/debug_tools.php';
 require_once __DIR__ . '/src/Admin/Settings.php';
 
 // Register event route (and any others)
-add_action('rest_api_init', 'VerticalAppDriver\Api\register_all_routes');
+add_action('rest_api_init', 'VerticalAppDriver\Api\Database\register_all_routes');
 
 // Activation hook: Set default settings
 register_activation_hook(__FILE__, function ()
@@ -34,6 +34,25 @@ register_activation_hook(__FILE__, function ()
 });
 
 add_action('admin_menu', 'VerticalAppDriver\Helpers\add_php_info_page');
+
+
+// Used to add hooks to events manager new event creation
+// (published for the first time or state changed from draft to published)
+add_filter('em_event_save', function($result, $EM_Event) {
+    if (!$result) return $result; // Save failed, ignore
+
+    // Get previous and current status
+    $previous_status = $EM_Event->get_previous_status(); // e.g. 'draft', 'pending', etc.
+    $current_status = $EM_Event->get_status(true); // true: get current status
+
+    // Only trigger when an event goes from not published to published
+    if ($previous_status !== '1' && $current_status === 1)
+    {
+        echo("New event detected !");
+        // This is a new publication or a status change to published
+    }
+    return $result;
+}, 10, 2);
 
 
 // Deactivation hook: Cleanup settings
